refactor: Diary 예외처리기능 리팩토링
