+++ src/main/java/homeTry/diary/exception/BadRequestException/MemoBlankException.java
@@ -1,5 +1,6 @@
-package homeTry.diary.exception;
+package homeTry.diary.exception.BadRequestException;
 
+import homeTry.diary.exception.DiaryErrorType;
 import homeTry.exception.BadRequestException;
 
 public class MemoBlankException extends BadRequestException{
